+++ src/main/ruby/truffleruby/core/enumerator.rb
@@ -400,7 +400,7 @@ class Enumerator
       taken = 0
       Lazy.new(self, set_size) do |yielder, *args|
         if taken < n
-          yielder.yield(*args)
+          yielder.yield(Truffle::EnumerableOperations.pack_values(args))
           taken += 1
           raise StopLazyError unless taken < n
         else
@@ -425,7 +425,7 @@ class Enumerator
         if dropped < n
           dropped += 1
         else
-          yielder.yield(*args)
+          yielder.yield(Truffle::EnumerableOperations.pack_values(args))
         end
       end
     end
@@ -435,7 +435,7 @@ class Enumerator
 
       Lazy.new(self, nil) do |yielder, *args|
         if yield(*args)
-          yielder.yield(*args)
+          yielder.yield(Truffle::EnumerableOperations.pack_values(args))
         else
           raise StopLazyError
         end
@@ -450,10 +450,10 @@ class Enumerator
         if succeeding
           unless yield(*args)
             succeeding = false
-            yielder.yield(*args)
+            yielder.yield(Truffle::EnumerableOperations.pack_values(args))
           end
         else
-          yielder.yield(*args)
+          yielder.yield(Truffle::EnumerableOperations.pack_values(args))
         end
       end
     end
@@ -471,8 +471,8 @@ class Enumerator
       raise ArgumentError, 'Lazy#{select,find_all} requires a block' unless block_given?
 
       Lazy.new(self, nil) do |yielder, *args|
-        val = args.length >= 2 ? args : args.first
-        yielder.yield(*args) if yield(val)
+        val = Truffle::EnumerableOperations.pack_values(args)
+        yielder.yield(val) if yield(val)
       end
     end
     alias_method :find_all, :select
@@ -482,8 +482,8 @@ class Enumerator
       raise ArgumentError, 'Lazy#reject requires a block' unless block_given?
 
       Lazy.new(self, nil) do |yielder, *args|
-        val = args.length >= 2 ? args : args.first
-        yielder.yield(*args) unless yield(val)
+        val = Truffle::EnumerableOperations.pack_values(args)
+        yielder.yield(val) unless yield(val)
       end
     end
 
@@ -492,7 +492,7 @@ class Enumerator
 
       Lazy.new(self, nil) do |yielder, *args|
         Primitive.share_special_variables(sv)
-        val = args.length >= 2 ? args : args.first
+        val = Truffle::EnumerableOperations.pack_values(args)
         matches = pattern === val
 
         if matches
@@ -509,7 +509,7 @@ class Enumerator
       s = block ? Primitive.proc_special_variables(block) : Primitive.caller_special_variables
 
       Lazy.new(self, nil) do |yielder, *args|
-        val = args.length >= 2 ? args : args.first
+        val = Truffle::EnumerableOperations.pack_values(args)
         matches = pattern === val
         Primitive.regexp_last_match_set(s, $~)
 
@@ -566,10 +566,11 @@ class Enumerator
 
       Lazy.new(self, enumerator_size) do |yielder, *args|
         memo = yielder.memo || offset
+        val = Truffle::EnumerableOperations.pack_values(args)
         if block
-          yielder.yield yield(*args, memo)
+          yielder.yield yield(val, memo)
         else
-          yielder.yield(*args, memo)
+          yielder.yield [val, memo]
         end
         yielder.memo = Primitive.rb_num2long(memo) + 1
       end
@@ -593,7 +594,7 @@ class Enumerator
 
       index = 0
       Lazy.new(self, enumerator_size) do |yielder, *args|
-        val = args.length >= 2 ? args : args.first
+        val = Truffle::EnumerableOperations.pack_values(args)
         rests = lists.map do |list|
           case list
           when Array
@@ -620,8 +621,9 @@ class Enumerator
     end
 
     def compact
-      Lazy.new(self, nil) do |yielder, e|
-        yielder.yield(e) unless Primitive.nil?(e)
+      Lazy.new(self, nil) do |yielder, *args|
+        val = Truffle::EnumerableOperations.pack_values(args)
+        yielder.yield(val) unless Primitive.nil?(val)
       end
     end
 
@@ -641,20 +643,20 @@ class Enumerator
       if block_given?
         h = {}
         Lazy.new(self, nil) do |yielder, *args|
-          val = args.length >= 2 ? args : args.first
+          val = Truffle::EnumerableOperations.pack_values(args)
           comp = yield(val)
           unless h.key?(comp)
             h[comp] = true
-            yielder.yield(*args)
+            yielder.yield(val)
           end
         end
       else
         h = {}
         Lazy.new(self, nil) do |yielder, *args|
-          val = args.length >= 2 ? args : args.first
+          val = Truffle::EnumerableOperations.pack_values(args)
           unless h.key?(val)
             h[val] = true
-            yielder.yield(*args)
+            yielder.yield(val)
           end
         end
       end